<VERIFY> [_8110] Correct narrowing of floating point values

+++ server/core/src/json_deserialization.cpp
@@ -77,12 +77,12 @@ namespace irods
                     p->inOutStruct = v;
                 }
                 else if (std::string_view{DOUBLE_MS_T} == p->type) {
-                    auto* v = static_cast<int*>(std::malloc(sizeof(double)));
+                    auto* v = static_cast<double*>(std::malloc(sizeof(double)));
                     *v = value.get<double>();
                     p->inOutStruct = v;
                 }
                 else if (std::string_view{FLOAT_MS_T} == p->type) {
-                    auto* v = static_cast<int*>(std::malloc(sizeof(float)));
+                    auto* v = static_cast<float*>(std::malloc(sizeof(float)));
                     *v = value.get<float>();
                     p->inOutStruct = v;
                 }
